fix html children problem, remove bottomItems and cloudBottomItems components
