--- airbyte-webapp/src/packages/cloud/views/layout/CloudMainView/CloudBottomItems.tsx
@@ -1,52 +1,0 @@
-import { FormattedMessage } from "react-intl";
-
-import { CreditsIcon } from "components/icons/CreditsIcon";
-
-import { FeatureItem, useFeature } from "hooks/services/Feature";
-import { CloudRoutes } from "packages/cloud/cloudRoutePaths";
-import { useGetCloudWorkspace } from "packages/cloud/services/workspaces/CloudWorkspacesService";
-import { LOW_BALANCE_CREDIT_THRESHOLD } from "packages/cloud/views/credits/CreditsPage/components/LowCreditBalanceHint/LowCreditBalanceHint";
-import { RoutePaths } from "pages/routePaths";
-import { useCurrentWorkspace } from "services/workspaces/WorkspacesService";
-import { MenuContent } from "views/layout/SideBar/components/MenuContent";
-import { NavItem } from "views/layout/SideBar/components/NavItem";
-import SettingsIcon from "views/layout/SideBar/components/SettingsIcon";
-
-import { CloudResourcesDropdown } from "./CloudResourcesDropdown";
-import { CloudSupportDropdown } from "./CloudSupportDropdown";
-
-export const CloudBottomItems: React.FC = () => {
-  const workspace = useCurrentWorkspace();
-  const cloudWorkspace = useGetCloudWorkspace(workspace.workspaceId);
-  const isAllowUpdateConnectorsEnabled = useFeature(FeatureItem.AllowUpdateConnectors);
-
-  return (
-    <ul>
-      <MenuContent>
-        <li>
-          <NavItem
-            to={CloudRoutes.Credits}
-            icon={<CreditsIcon />}
-            label={<FormattedMessage id="sidebar.credits" />}
-            testId="creditsButton"
-            withNotification={cloudWorkspace.remainingCredits <= LOW_BALANCE_CREDIT_THRESHOLD}
-          />
-        </li>
-        <li>
-          <CloudResourcesDropdown />
-        </li>
-        <li>
-          <CloudSupportDropdown />
-        </li>
-        <li>
-          <NavItem
-            label={<FormattedMessage id="sidebar.settings" />}
-            icon={<SettingsIcon />}
-            to={RoutePaths.Settings}
-            withNotification={isAllowUpdateConnectorsEnabled}
-          />
-        </li>
-      </MenuContent>
-    </ul>
-  );
-};
