+++ airbyte-webapp/src/views/layout/MainView/MainView.tsx
@@ -1,16 +1,24 @@
+import { faRocket } from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import React from "react";
+import { FormattedMessage } from "react-intl";
 
 import { LoadingPage } from "components";
+import { Version } from "components/common/Version";
 import { FlexContainer } from "components/ui/Flex";
 
 import { useConfig } from "config";
+import { RoutePaths } from "pages/routePaths";
+import { links } from "utils/links";
 import { ResourceNotFoundErrorBoundary } from "views/common/ResorceNotFoundErrorBoundary";
 import { StartOverErrorView } from "views/common/StartOverErrorView";
 
 import styles from "./MainView.module.scss";
 import { AirbyteHomeLink } from "../SideBar/AirbyteHomeLink";
-import { BottomItems } from "../SideBar/BottomItems";
 import { MenuContent } from "../SideBar/components/MenuContent";
+import { NavItem } from "../SideBar/components/NavItem";
+import { ResourcesDropdown } from "../SideBar/components/ResourcesDropdown";
+import SettingsIcon from "../SideBar/components/SettingsIcon";
 import { MainNavItems } from "../SideBar/MainNavItems";
 import { SideBar } from "../SideBar/SideBar";
 
@@ -23,7 +31,23 @@ const MainView: React.FC<React.PropsWithChildren<unknown>> = (props) => {
         <AirbyteHomeLink />
         <MenuContent>
           <MainNavItems />
-          <BottomItems version={version} />
+          <MenuContent>
+            <NavItem
+              as="a"
+              to={links.updateLink}
+              icon={<FontAwesomeIcon icon={faRocket} />}
+              label={<FormattedMessage id="sidebar.update" />}
+              testId="updateLink"
+            />
+            <ResourcesDropdown />
+            <NavItem
+              label={<FormattedMessage id="sidebar.settings" />}
+              icon={<SettingsIcon />}
+              to={RoutePaths.Settings}
+              withNotification
+            />
+            {version && <Version primary />}
+          </MenuContent>
         </MenuContent>
       </SideBar>
       <div className={styles.content}>
