+++ airbyte-webapp/src/views/layout/SideBar/SideBar.tsx
@@ -1,12 +1,8 @@
+import classNames from "classnames";
 import { PropsWithChildren } from "react";
 
-import { MenuContent } from "./components/MenuContent";
 import styles from "./SideBar.module.scss";
 
 export const SideBar: React.FC<PropsWithChildren<unknown>> = ({ children }) => {
-  return (
-    <nav className={styles.nav}>
-      <MenuContent>{children}</MenuContent>
-    </nav>
-  );
+  return <nav className={classNames(styles.nav)}>{children}</nav>;
 };
