+++ airbyte-webapp/src/views/layout/SideBar/components/MenuContent.module.scss
@@ -2,4 +2,8 @@
   height: 100%;
   padding: 0;
   width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
 }
