+++ airbyte-webapp/src/views/layout/SideBar/components/MenuContent.tsx
@@ -1,19 +1,13 @@
 import { PropsWithChildren } from "react";
 
-import { FlexContainer } from "components/ui/Flex";
-
 import styles from "./MenuContent.module.scss";
 
-export const MenuContent: React.FC<PropsWithChildren<unknown>> = ({ children }) => {
+export const MenuContent: React.FC<PropsWithChildren<{ children: React.ReactNode[] }>> = ({ children }) => {
   return (
-    <FlexContainer
-      direction="column"
-      gap="xs"
-      alignItems="center"
-      justifyContent="space-between"
-      className={styles.menuContent}
-    >
-      {children}
-    </FlexContainer>
+    <ul className={styles.menuContent}>
+      {children.map((child, index) => {
+        return <li key={index}>{child}</li>;
+      })}
+    </ul>
   );
 };
